Replaces viewport module with more useful mixin
Also removes some non-useful JS
