--- resources/assets/js/modules/bem.js
@@ -1,6 +1,0 @@
-export const states = {
-	active: 'is-active',
-	fixed: 'is-fixed',
-	hidden: 'is-hidden',
-	loading: 'is-loading',
-};
